getDatabaseOptions _util
